only superusers can edit add and delete products

+++ products/views.py
@@ -3,6 +3,8 @@ from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
 from .forms import ProductForm
+from django.contrib.auth.decorators import login_required
+
 
 # Create your views here.
 
@@ -67,8 +69,13 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
+
+@login_required
 def add_product(request):
     """ allows admin to add products to the store """
+    if not request.user.is_superuser:
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -84,8 +91,13 @@ def add_product(request):
 
     return render(request, template, context)
 
+
+@login_required
 def edit_product(request, product_id):
     """ allows admin to edit products already created """
+    if not request.user.is_superuser:
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         #informing im editing the product defined above (instance=product)
@@ -105,8 +117,12 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ allows admin to delete a product """
+    if not request.user.is_superuser:
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     return redirect(reverse('products'))
